requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
requests==2.32.3
backoff==2.2.1
psycopg2-binary==2.9.9
